Do not remove loggers during Clear that fail to close properly
Separate gRPC service into file based on interface

Return special error code if logger is not found in Remove
