+++ remote/internal/service.go
@@ -8,9 +8,7 @@ import (
 	"time"
 
 	"github.com/evergreen-ci/lru"
-	"github.com/golang/protobuf/ptypes"
 	empty "github.com/golang/protobuf/ptypes/empty"
-	timestamp "github.com/golang/protobuf/ptypes/timestamp"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/recovery"
 	"github.com/mongodb/jasper"
@@ -547,230 +545,6 @@ func (s *jasperService) ScriptingHarnessGet(ctx context.Context, id *ScriptingHa
 	return &OperationOutcome{
 		Success: true,
 		Text:    sh.ID(),
-	}, nil
-}
-
-func (s *jasperService) ScriptingHarnessSetup(ctx context.Context, id *ScriptingHarnessID) (*OperationOutcome, error) {
-	sh, err := s.scripting.Get(id.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Wrapf(err, "getting scripting harness with id '%s'", id.Id))
-	}
-
-	if err = sh.Setup(ctx); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "setting up scripting harness"))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) ScriptingHarnessCleanup(ctx context.Context, id *ScriptingHarnessID) (*OperationOutcome, error) {
-	sh, err := s.scripting.Get(id.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Wrapf(err, "getting scripting harness with id '%s'", id.Id))
-	}
-
-	if err = sh.Cleanup(ctx); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "cleaning up scripting harness"))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) ScriptingHarnessRun(ctx context.Context, args *ScriptingHarnessRunArgs) (*OperationOutcome, error) {
-	sh, err := s.scripting.Get(args.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Wrapf(err, "getting scripting harness with id '%s'", args.Id))
-	}
-
-	if err = sh.Run(ctx, args.Args); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "running scripting command"))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) ScriptingHarnessRunScript(ctx context.Context, args *ScriptingHarnessRunScriptArgs) (*OperationOutcome, error) {
-	sh, err := s.scripting.Get(args.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Wrapf(err, "getting scripting harness with id '%s'", args.Id))
-	}
-
-	err = sh.RunScript(ctx, args.Script)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "running script"))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) ScriptingHarnessBuild(ctx context.Context, args *ScriptingHarnessBuildArgs) (*ScriptingHarnessBuildResponse, error) {
-	sh, err := s.scripting.Get(args.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Wrapf(err, "getting scripting harness with id '%s'", args.Id))
-	}
-
-	path, err := sh.Build(ctx, args.Directory, args.Args)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "running build"))
-	}
-
-	return &ScriptingHarnessBuildResponse{
-		Path: path,
-		Outcome: &OperationOutcome{
-			Success: true,
-		}}, nil
-}
-
-func (s *jasperService) ScriptingHarnessTest(ctx context.Context, args *ScriptingHarnessTestArgs) (*ScriptingHarnessTestResponse, error) {
-	sh, err := s.scripting.Get(args.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Wrapf(err, "getting scripting harness with id '%s'", args.Id))
-	}
-
-	exportedArgs, err := args.Export()
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "exporting arguments"))
-	}
-
-	var testErr error
-	res, err := sh.Test(ctx, args.Directory, exportedArgs...)
-	if err != nil {
-		testErr = err
-	}
-	convertedRes, err := ConvertScriptingTestResults(res)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "converting test results"))
-	}
-
-	outcome := &OperationOutcome{Success: testErr == nil}
-	if testErr != nil {
-		outcome.Text = testErr.Error()
-	}
-	return &ScriptingHarnessTestResponse{
-		Outcome: outcome,
-		Results: convertedRes,
-	}, nil
-}
-
-var errLoggingCacheNotSupported = errors.New("logging cache is not supported")
-
-func (s *jasperService) LoggingCacheCreate(ctx context.Context, args *LoggingCacheCreateArgs) (*LoggingCacheInstance, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-	opts, err := args.Options.Export()
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "exporting options"))
-	}
-	if err := opts.Validate(); err != nil {
-		return nil, newGRPCError(codes.InvalidArgument, errors.Wrap(err, "invalid options"))
-	}
-
-	out, err := lc.Create(args.Id, &opts)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "creating logger"))
-	}
-	out.ManagerID = s.manager.ID()
-
-	logger, err := ConvertCachedLogger(out)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "converting cached logger"))
-	}
-
-	return logger, nil
-}
-
-func (s *jasperService) LoggingCacheGet(ctx context.Context, args *LoggingCacheArgs) (*LoggingCacheInstance, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-
-	out, err := lc.Get(args.Id)
-	if err != nil {
-		return nil, newGRPCError(codes.NotFound, errors.Errorf("getting logger with id '%s'", args.Id))
-	}
-
-	logger, err := ConvertCachedLogger(out)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "converting cached logger"))
-	}
-
-	return logger, nil
-}
-
-func (s *jasperService) LoggingCacheRemove(ctx context.Context, args *LoggingCacheArgs) (*OperationOutcome, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-
-	if err := lc.Remove(args.Id); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "removing logger with id '%s'", args.Id))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) LoggingCacheCloseAndRemove(ctx context.Context, args *LoggingCacheArgs) (*OperationOutcome, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-
-	if err := lc.CloseAndRemove(ctx, args.Id); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrapf(err, "closing and removing logger with id '%s'", args.Id))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) LoggingCacheClear(ctx context.Context, _ *empty.Empty) (*OperationOutcome, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-
-	if err := lc.Clear(ctx); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "clearing logging cache"))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) LoggingCachePrune(ctx context.Context, arg *timestamp.Timestamp) (*OperationOutcome, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-
-	ts, err := ptypes.Timestamp(arg)
-	if err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "converting input timestamp"))
-	}
-
-	if err := lc.Prune(ts); err != nil {
-		return nil, newGRPCError(codes.Internal, errors.Wrap(err, "pruning logging cache"))
-	}
-
-	return &OperationOutcome{Success: true}, nil
-}
-
-func (s *jasperService) LoggingCacheLen(ctx context.Context, _ *empty.Empty) (*LoggingCacheLenResponse, error) {
-	lc := s.manager.LoggingCache(ctx)
-	if lc == nil {
-		return nil, newGRPCError(codes.FailedPrecondition, errLoggingCacheNotSupported)
-	}
-
-	length, err := lc.Len()
-	if err != nil {
-		return nil, errors.Wrap(err, "getting logging cache length")
-	}
-
-	return &LoggingCacheLenResponse{
-		Outcome: &OperationOutcome{Success: true},
-		Len:     int64(length),
 	}, nil
 }
